Moe/Ursula Status App: move package to be under network
