+++ nucypher/network/status/assets/favicon.ico
