+++ nucypher/network/status/assets/nucypher_logo.png
